Clarify lzma memlimit
Note the memlimit parameter is a bit confusing. It sounds like it might
convince liblzma to switch to a lower-memory algorithm or something, but
it doesn't do that. Instead it just makes liblzma fail-fast if it would
have used too much memory.

I'm not sure there's real value in setting this. After all, we don't set it
for compression (which uses much more memory), nor for any of the other
decompression methods.

But 65 MB is the max value documented in xz(1)[0], and is low enough that
any vaguely modern computer can handle it, so there's no harm. The only minor
concern is that someday liblzma changes its memory requirements, and we'll
have to raise this.

[0] https://tukaani.org/xz/man/xz.1.html#DESCRIPTION:_Memory_usage